Highlighted text is appended
When text is highlighted and "Ask Assistant" is used, it will be appended to the text box in the side panel.

Console logs are added to see if the side panel is currently open or text is selected for debugging purposes.

Removed ⏳ icon that appears when text is selected and is being sent to the text box.